pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "navidiv"
maintainers = [
  { name = "Mohammed Azzouzi", email = "mohammed.azzouzi@epfl.ch" },
]

dependencies = [
  "numpy",
  "scipy",
  "pandas",
  "matplotlib",
  "rdkit",
  "streamlit",
  "scikit-learn",
  "igraph",
  "hydra-core",
]
requires-python = ">=3.12"
dynamic = ["version"]
readme = "README.rst"
description = "Does things for you."

[dependency-groups]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "pytest-cov",
  "sphinx",
  "sphinx-copybutton",
  "furo",
]

[project.urls]
github = "https://github.com/mohammedazzouzi15/template"

[tool.setuptools_scm]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "ANN401", "COM812", "ISC001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
  "S101",
  "INP001",
]
"examples/*" = [
  "INP001",
]
"docs/source/conf.py" = ["D100", "INP001"]

[tool.mypy]
show_error_codes = true
implicit_optional = false
warn_no_return = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unreachable = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report term-missing"
